[1]: 添加了日程表已完成勾选功能、基本了日程表对应后端逻辑、美化了用户已完成卡片外形及标签

+++ src/main/java/com/origin/entity/Schedule.java
@@ -13,7 +13,7 @@ public class Schedule {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id; // 必须有主键
+    private Integer id;
 
     @Column(name = "name", nullable = false)
     private String name;
